Move event cache into own submodule
